pyproject.toml
[tool.poetry]
name = "ca-certificates"
version = "2"
description = "Utilities for system wide CA certificate installation"
authors = ["Ludwig Nussel <ludwig.nussel@suse.de>"]
license = "GPL-2.0"

[tool.poetry.dependencies]
python = ">= 3.6.2"

[tool.poetry.dev-dependencies]
pytest-testinfra = "^6.3.0"
pytest = "^6.2.4"
black = "^21.5b1"
pytest-xdist = "^2.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py36']

[tool.pytest.ini_options]
xfail_strict = true
